% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_select_candidate.R
\name{greedy_select_candidate}
\alias{greedy_select_candidate}
\title{Select candidate areas using greedy selection algorithm}
\usage{
greedy_select_candidate(
  goal_vec,
  candidate_mat,
  cv_schemes,
  n_select = 10,
  rcpp = TRUE
)
}
\arguments{
\item{goal_vec}{Vector of time-stratified (e.g., weekly) cases in the target area}

\item{candidate_mat}{Matrix of candidate cases (time units × candidate areas)}

\item{cv_schemes}{List of length \code{n_cv_schemes}, each element is a vector of block labels (e.g., "A", "B", "C") of length \code{tot_ts}. Output of \code{make_cv_schemes()}}

\item{n_select}{Number of candidates to select}

\item{rcpp}{Whether to use Rcpp compiler function or R function to evaluate candidates}
}
\value{
A list with:
\itemize{
  \item \code{selected_areas}: Vector of \code{n_select} selected areas
  \item \code{corr2_per_step}: Vector of performance with each additional selection
}
}
\description{
This function selects candidate areas using a greedy selection algorithm by comparing the out-of-sample corr² averaged across test blocks and cross-validation schemes
}
