% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_select.R
\name{order_select}
\alias{order_select}
\title{Count number of times candidate areas are selected}
\usage{
order_select(mat_sel)
}
\arguments{
\item{mat_sel}{Matrix of selected areas. Rows represent selections in each iterations}
}
\value{
A data frame with columns of area name and the frequency of selection, ordered by the most commonly-selected areas
}
\description{
This function counts the number of times a candidate area is selected across all iterations
}
