% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_targ.R
\name{pred_targ}
\alias{pred_targ}
\title{Predict target area from selected areas}
\usage{
pred_targ(goal_vec, mat_sel)
}
\arguments{
\item{goal_vec}{Vector of time-stratified (e.g., weekly) cases in the target area}

\item{mat_sel}{Matrix of selected areas. Rows represent selections in each iterations}
}
\value{
A data frame with columns of area name and the frequency of selection, ordered by the most commonly-selected areas
#' @return A list with:
\itemize{
  \item \code{pred}: Vector of time-series of predicted values for target area
  \item \code{corr2}: Corr² value between predicted and target
}
}
\description{
This function runs multivariate linear regression on time-series of values from selected areas to predict the time-series of values from target areas
}
